pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "myl"
authors = [
  {name = "Philipp Schmitt", email = "philipp@schmitt.co"},
]
description = "Dead simple IMAP CLI client"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["imap", "email"]
license = {file = "LICENSE"}
classifiers = [
  "Programming Language :: Python :: 3",
]
dependencies = [
  "dnspython",
  "imap_tools",
  "requests",
  "rich",
  "xmltodict"
]
version = "0.7.1"

[tool.black]
line-length = 79

[project.scripts]
myl = "myl:main"
